moving request and response to the correct file
