+++ ironfish/src/rpc/routes/wallet/create.ts
@@ -9,31 +9,9 @@
  * object being acted upon. For example, `POST /wallet/burnAsset` burns an asset.
  */
 
-import * as yup from 'yup'
 import { ERROR_CODES, ValidationError } from '../../adapters'
 import { ApiNamespace, routes } from '../router'
-
-export type CreateAccountRequest = { name: string; default?: boolean }
-export type CreateAccountResponse = {
-  name: string
-  publicAddress: string
-  isDefaultAccount: boolean
-}
-
-export const CreateAccountRequestSchema: yup.ObjectSchema<CreateAccountRequest> = yup
-  .object({
-    name: yup.string().defined(),
-    default: yup.boolean().optional(),
-  })
-  .defined()
-
-export const CreateAccountResponseSchema: yup.ObjectSchema<CreateAccountResponse> = yup
-  .object({
-    name: yup.string().defined(),
-    publicAddress: yup.string().defined(),
-    isDefaultAccount: yup.boolean().defined(),
-  })
-  .defined()
+import { CreateAccountRequestSchema, CreateAccountResponse } from '../wallet'
 
 routes.register<typeof CreateAccountRequestSchema, CreateAccountResponse>(
   `${ApiNamespace.wallet}/create`,
